stock 서비스 로직 작성 및 h2 데이터베이스 스키마, 샘플 데이터 수정

+++ src/main/java/com/example/gymerp/dto/ServiceDto.java
@@ -1,6 +1,6 @@
 package com.example.gymerp.dto;
 
-import java.math.BigDecimal;
+
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -24,7 +24,7 @@ public class ServiceDto {
 	private String codeA;
 	private String codeB;
 	private String name;
-	private BigDecimal price;
+	private int price;
 	private Boolean isActive;
 	private String note;
 	private int serviceValue;
